+++ quilt-tester/tests/build-dockerfile/build-dockerfile.js
@@ -12,8 +12,6 @@ for (let i = 0; i < infrastructure.nWorker; i++) {
                 'RUN echo ' + i + ' > /dockerfile-id\n' +
                 'RUN echo $(cat /dev/urandom | tr -dc \'a-zA-Z0-9\' | ' +
                 'fold -w 32 | head -n 1) > /image-id'
-            ),
-            ['tail', '-f', '/dev/null']
-        ).replicate(2)
+            ), {command: ['tail', '-f', '/dev/null']}).replicate(2)
     ));
 }