+++ quilt-tester/tests/outbound-public/outbound-public.js
@@ -5,13 +5,13 @@ let deployment = quilt.createDeployment();
 deployment.deploy(infrastructure);
 
 let connected = new quilt.Service('connected',
-    new quilt.Container('alpine', ['tail', '-f', '/dev/null'])
+    new quilt.Container('alpine', {command: ['tail', '-f', '/dev/null']})
         .replicate(infrastructure.nWorker*2)
 );
 quilt.publicInternet.allowFrom(connected, 80);
 
 let notConnected = new quilt.Service('not-connected',
-    new quilt.Container('alpine', ['tail', '-f', '/dev/null'])
+    new quilt.Container('alpine', {command: ['tail', '-f', '/dev/null']})
         .replicate(infrastructure.nWorker*2)
 );